Moved CategoryBreakdown to ProgressScreen folder
